+++ create-vtj/templates/h5/src/main.ts
@@ -8,7 +8,8 @@ import {
   notify,
   loading,
   createAdapter,
-  createServiceRequest
+  createServiceRequest,
+  Startup
 } from '@vtj/h5';
 import { createApp } from 'vue';
 import router from './router';
@@ -17,7 +18,7 @@ import { name } from '../package.json';
 import './style/index.scss';
 
 const app = createApp(App);
-const adapter = createAdapter({ loading, notify });
+const adapter = createAdapter({ loading, notify, Startup });
 const service = new LocalService(createServiceRequest(notify));
 const { provider, onReady } = createProvider({
   nodeEnv: process.env.NODE_ENV as NodeEnv,