+++ packages/coder/src/templates.ts
@@ -4,7 +4,7 @@ const scriptTemplate = `
 // @ts-nocheck
 
 <%= imports %>
-import { useProvider } from '@vtj/renderer';
+import { useProvider } from '<%= renderer %>';
 export default defineComponent({
   name: '<%= name %>', 
   <% if(inject) { %> inject: { <%= inject %>}, <% } %>