+++ packages/renderer/src/provider/startup.ts
@@ -1,8 +1,8 @@
 import { defineComponent, h } from 'vue';
-import { Startup } from '../components';
+// import { Startup } from '../components';
 export const StartupContainer = defineComponent({
   name: 'VtjStartupContainer',
   render() {
-    return h(Startup);
+    return h('div', 'page not found!');
   }
 });