+++ platforms/h5/src/components/index.ts
@@ -1,5 +1,3 @@
 import Startup from './Startup.vue';
-export * from '@vtj/ui';
-export * from '@vtj/charts';
 
 export { Startup };