+++ platforms/uni-app/src/modules.ts
@@ -4,10 +4,9 @@ export function createModules(mode: ContextMode = ContextMode.Runtime) {
   if (mode === ContextMode.Runtime || process.env.NODE_ENV === 'development') {
     return import.meta.glob([
       '/src/.vtj/projects/*.json',
-      '/src/.vtj/files/*.json',
-      '/src/pages/*.vue'
+      '/src/.vtj/files/*.json'
     ]);
   } else {
-    return import.meta.glob(['/src/.vtj/projects/*.json', '/src/pages/*.vue']);
+    return import.meta.glob(['/src/.vtj/projects/*.json']);
   }
 }